pe-dashboard-ai50-v3-main/requirements.txt
pydantic>=2.7.0
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.27.0
langchain>=0.2.0
langgraph>=0.1.0
openai>=1.35.0,<2
chromadb
langchain-core
langchain-text-splitters
langchain-openai
python-dotenv==1.0.1
pytest>=7.0.0
pytest-asyncio>=0.21.0
feedparser>=6.0.0
requests>=2.31.0
pendulum>=3.0.0
beautifulsoup4>=4.12.0
google-cloud-storage
tabulate
streamlit